+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+- Added check if commands available in the system.
+- automatically install missing commands.
+
 ## 1.0.0
 
 - Initial version.